added basic player movement and camera panning
In this push i add placeholder sprites for the player and floor as well as the objects for the player, floor, and camera. the player has basic movement controls and the camera pans to keep the player centered. i also made the game go full screen on start and allowed players to exit the game with escape.